backend/pyproject.toml
[project]
name = "BCI"
version = "0.1.0"
description = "BCI for Thonk"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "cloudpickle>=3.0",
    "fsspec>=2024.10",
    "h5py>=3.10",
    "mat73>=0.60",
    "matplotlib>=3.8",
    "numpy>=1.26.4",
    "pandas>=2.1",
    "protobuf>=4.25",
    "PyYAML>=6.0",
    "scikit-image>=0.22",
    "scikit-learn>=1.4",
    "scipy>=1.11",
    "tqdm>=4.66",
    
    "mediapipe>=0.10.1",
    "opencv-python>=4.9",
    "pycocotools>=2.0.7",
    "imageio>=2.34.0",
    "imageio-ffmpeg>=0.5.0",
    
    "dspy>=1.4.2",
    "einops>=0.8.1",
    "fvcore>=0.1.5",
    "huggingface_hub>=0.26.0",
    "omegaconf>=2.3.0",
    "tensorboardX>=2.6",
    "tensorflow>=2.16",
    "thop>=0.1.1",
    "timm>=1.0",
    "torch>=2.2",
    "torchaudio>=2.1",
    "torchvision>=0.16",
    "yacs>=0.1.8",
    "transformers>=4.37.0",
    "mne>=1.5.0",
    
    "brainflow>=5.19.0",

    "neurokit2>=0.2.10",
    "pyproj>=3.7.1",
    
    "accelerate>=0.26.0",
    "python-multipart>=0.0.9",

    "fastapi>=0.115.12",
    "gradio>=3.40.1",
    "requests>=2.31.0",
    "uvicorn>=0.24.0",
    "websockets>=11.0.3",
    
    "mypy>=1.18.2",
    "pytest>=7.4.2",
    "python-dotenv>=1.1.0",
    "setuptools<81",
]
